+++ setup.py
@@ -1,11 +1,7 @@
 #!/usr/bin/python
 
 __author__ = 'Ryan Cox <ryan.a.cox@gmail.com>'
-__version__ = '1.2'
-
-test_requirements = [
-    # TODO: put package test requirements here
-]
+__version__ = '1.3'
 
 # Distutils version
 METADATA = dict(
@@ -19,21 +15,26 @@ METADATA = dict(
     url = 'http://github.com/ryancox/motionless',
     keywords = 'google static maps url api georss mapping gpx kml geo gis',
     test_suite='tests',
-    tests_require=test_requirements,
+    tests_require=[],
 )
 
 # Setuptools version
 SETUPTOOLS_METADATA = dict(
-    install_requires = ['setuptools'],
+    install_requires = ['setuptools', 'six'],
     include_package_data = True,
     classifiers = [
         'Development Status :: 4 - Beta',
         'Intended Audience :: Developers',
         'License :: OSI Approved :: Apache Software License',
         'Topic :: Software Development :: Libraries :: Python Modules',
-                'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
-                'Topic :: Multimedia :: Graphics :: Presentation',
+        'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
+        'Topic :: Multimedia :: Graphics :: Presentation',
         'Topic :: Internet',
+        'Programming Language :: Python :: 2',
+        'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.4',
+        'Programming Language :: Python :: 3.5',
     ]
 )
 
@@ -47,4 +48,4 @@ def Main():
         distutils.core.setup(**METADATA)
 
 if __name__ == '__main__':
-  Main()
+    Main()
